+++ rmf_traffic/include/rmf_traffic/geometry/Shape.hpp
@@ -115,20 +115,6 @@ FinalShapePtr make_final(const T& shape)
   return std::make_shared<FinalShape>(shape.finalize());
 }
 
-//==============================================================================
-template<typename T>
-std::function<bool(const Shape& other)> make_equality_comparator(const T& myself)
-{
-  return [&myself](const Shape& other)
-  {
-    if (const auto* other_derived = dynamic_cast<const T*>(&other))
-    {
-      return myself == *other_derived;
-    }
-    return false;
-  };
-}
-
 } // namespace geometry
 } // namespace rmf_traffic