+++ rmf_traffic/include/rmf_traffic/schedule/Mirror.hpp
@@ -74,10 +74,14 @@ public:
   /// Create a database mirror
   Mirror();
 
+  /// Update the known participants and their descriptions.
+  void update_participants_info(const ParticipantDescriptionsMap& participants);
+
   /// Update this mirror.
   ///
-  /// \return the last version that this Mirror knows of
-  Version update(const Patch& patch);
+  /// \return true if this update is okay. false if the base version of the
+  /// patch does not match
+  bool update(const Patch& patch);
 
   // TODO(MXG): Consider a feature to log and report any possible
   // inconsistencies that might show up with the patches, e.g. replacing or