+++ rmf_traffic/src/rmf_traffic/geometry/ShapeInternal.hpp
@@ -114,6 +114,21 @@ public:
   }
 };
 
+//==============================================================================
+template<typename T>
+std::function<bool(const Shape& other)> make_equality_comparator(
+  const T& myself)
+{
+  return [&myself](const Shape& other)
+    {
+      if (const auto* other_derived = dynamic_cast<const T*>(&other))
+      {
+        return myself == *other_derived;
+      }
+      return false;
+    };
+}
+
 } // namespace geometry
 } // namespace rmf_traffic